DataPanelUpdate

+++ Assets/Scripts/PublicVariables.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class PublicVariables : MonoBehaviour
 {
     //Player Health
-    public static float player_currentHealth = 5;
-    public static float player_maxHealth = 5;
+    public static int player_currentHealth = 5;
+    public static int player_maxHealth = 5;
 
     //Player Dash
     public static bool dash_enable = true;
